+++ packages/intl-locale/polyfill-force.ts
@@ -1,0 +1,8 @@
+import {Locale} from './';
+
+Object.defineProperty(Intl, 'Locale', {
+  value: Locale,
+  writable: true,
+  enumerable: false,
+  configurable: true,
+});