+++ packages/intl-numberformat/src/format_to_parts.ts
@@ -1,5 +1,6 @@
 import {
   DecimalFormatNum,
+  LDMLPluralRule,
   LDMLPluralRuleMap,
   NumberFormatLocaleInternalData,
   RawNumberFormatResult,
@@ -520,5 +521,5 @@ function selectPlural<T>(
   x: number,
   rules: LDMLPluralRuleMap<T>
 ): T {
-  return rules[pl.select(x) as Intl.LDMLPluralRule] || rules.other;
+  return rules[pl.select(x) as LDMLPluralRule] || rules.other;
 }