+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.6](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.10.5...@formatjs/cli@1.10.6) (2020-07-17)
+
+**Note:** Version bump only for package @formatjs/cli
+
+
+
+
+
+## [1.10.5](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.10.4...@formatjs/cli@1.10.5) (2020-07-16)
+
+**Note:** Version bump only for package @formatjs/cli
+
+
+
+
+
+## [1.10.4](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.10.3...@formatjs/cli@1.10.4) (2020-07-15)
+
+**Note:** Version bump only for package @formatjs/cli
+
+
+
+
+
+## [1.10.3](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.10.1...@formatjs/cli@1.10.3) (2020-07-14)
+
+**Note:** Version bump only for package @formatjs/cli
+
+
+
+
+
+## [1.10.2](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.10.1...@formatjs/cli@1.10.2) (2020-07-14)
+
+**Note:** Version bump only for package @formatjs/cli
+
+
+
+
+
+## [1.10.1](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.10.0...@formatjs/cli@1.10.1) (2020-07-14)
+
+**Note:** Version bump only for package @formatjs/cli
+
+
+
+
+
 # [1.10.0](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.9.3...@formatjs/cli@1.10.0) (2020-07-14)