+++ packages/intl-messageformat/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.1.5](https://github.com/formatjs/formatjs/compare/intl-messageformat@9.1.4...intl-messageformat@9.1.5) (2020-07-17)
+
+
+### Bug Fixes
+
+* **intl-messageformat:** add back intl-messageformat.umd.js, fix [#1846](https://github.com/formatjs/formatjs/issues/1846) ([81465b0](https://github.com/formatjs/formatjs/commit/81465b09bbaf44957d13d4e0f4ace23a2030047d))
+
+
+
+
+
+## [9.1.4](https://github.com/formatjs/formatjs/compare/intl-messageformat@9.1.3...intl-messageformat@9.1.4) (2020-07-16)
+
+**Note:** Version bump only for package intl-messageformat
+
+
+
+
+
+## [9.1.3](https://github.com/formatjs/formatjs/compare/intl-messageformat@9.1.1...intl-messageformat@9.1.3) (2020-07-14)
+
+
+### Bug Fixes
+
+* **react-intl:** fix rollup'ed type def file ([bddb88e](https://github.com/formatjs/formatjs/commit/bddb88e7435854b3152f2fbdc72b50054d9bad76))
+
+
+
+
+
+## [9.1.2](https://github.com/formatjs/formatjs/compare/intl-messageformat@9.1.1...intl-messageformat@9.1.2) (2020-07-14)
+
+
+### Bug Fixes
+
+* **react-intl:** fix rollup'ed type def file ([bddb88e](https://github.com/formatjs/formatjs/commit/bddb88e7435854b3152f2fbdc72b50054d9bad76))
+
+
+
+
+
+## [9.1.1](https://github.com/formatjs/formatjs/compare/intl-messageformat@9.1.0...intl-messageformat@9.1.1) (2020-07-14)
+
+**Note:** Version bump only for package intl-messageformat
+
+
+
+
+
 # [9.1.0](https://github.com/formatjs/formatjs/compare/intl-messageformat@9.0.2...intl-messageformat@9.1.0) (2020-07-14)