+++ packages/intl-numberformat/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.1](https://github.com/formatjs/formatjs/compare/@formatjs/intl-numberformat@5.3.0...@formatjs/intl-numberformat@5.3.1) (2020-07-16)
+
+
+### Reverts
+
+* Revert "feat(@formatjs/intl-utils): remove custom LDMLPluralRuleType and use TS3.9 type" ([1a6eeac](https://github.com/formatjs/formatjs/commit/1a6eeac8c59184825968ffddc35374c83f3fe782))
+
+
+
+
+
+# [5.3.0](https://github.com/formatjs/formatjs/compare/@formatjs/intl-numberformat@5.2.0...@formatjs/intl-numberformat@5.3.0) (2020-07-14)
+
+
+### Bug Fixes
+
+* **react-intl:** fix rollup'ed type def file ([bddb88e](https://github.com/formatjs/formatjs/commit/bddb88e7435854b3152f2fbdc72b50054d9bad76))
+
+
+### Features
+
+* **@formatjs/intl-numberformat:** add new digit mapping per spec, fix [#1801](https://github.com/formatjs/formatjs/issues/1801) ([aab965d](https://github.com/formatjs/formatjs/commit/aab965dd7c3a31da920122bfbe4e3069ad907ab8))
+
+
+
+
+
+## [5.2.1](https://github.com/formatjs/formatjs/compare/@formatjs/intl-numberformat@5.2.0...@formatjs/intl-numberformat@5.2.1) (2020-07-14)
+
+
+### Bug Fixes
+
+* **react-intl:** fix rollup'ed type def file ([bddb88e](https://github.com/formatjs/formatjs/commit/bddb88e7435854b3152f2fbdc72b50054d9bad76))
+
+
+
+
+
+# [5.2.0](https://github.com/formatjs/formatjs/compare/@formatjs/intl-numberformat@5.1.0...@formatjs/intl-numberformat@5.2.0) (2020-07-14)
+
+
+### Features
+
+* **@formatjs/intl-utils:** remove custom LDMLPluralRuleType and use TS3.9 type ([8e433d5](https://github.com/formatjs/formatjs/commit/8e433d5afdbf64f7a952d6345e50cad29bbb7083))
+
+
+
+
+
 # [5.1.0](https://github.com/formatjs/formatjs/compare/@formatjs/intl-numberformat@5.0.2...@formatjs/intl-numberformat@5.1.0) (2020-07-14)